minor change - refactoring
